[Change] separate author controllers from routes
