+++ routes/authorRoutes.js
@@ -1,77 +1,16 @@
 const express = require("express");
 
-const { dbPoolPromise } = require("../mySqlConnection");
 const { verifyJWT } = require("../JWT/jwtMiddleware");
 const { catchRequestError } = require("../utils/catchAsync");
+const author = require("../controllers/author_controllers");
 
 const router = express.Router();
 
-router.route("/").get(
-    catchRequestError(async (req, res) => {
-        const [authors] = await dbPoolPromise.execute("SELECT * FROM authors");
-        res.send(authors);
-    })
-);
+router.route("/").get(catchRequestError(author.findAll));
 router
     .route("/:id")
-    .post(
-        verifyJWT,
-        catchRequestError(async (req, res) => {
-            const id = req.body.id;
-
-            const [author] = await dbPoolPromise.execute(
-                "SELECT * FROM authors WHERE id = ?",
-                [id]
-            );
-            res.send(author);
-        })
-    )
-    .put(
-        verifyJWT,
-        catchRequestError(async (req, res) => {
-            first_name = req.body.name;
-            last_name = req.body.last_name;
-            images = req.body.images;
-            url = req.body.url;
-            bio = req.body.bio;
-            id = req.body.id;
-
-            const tempImgs = JSON.stringify(images);
-
-            await dbPoolPromise.execute(
-                "UPDATE authors SET name = ?, last_name =?, img =?, url =?, bio =? WHERE id =?",
-                [first_name, last_name, tempImgs, url, bio, id]
-            );
-
-            res.send("author edited");
-        })
-    )
-    .delete(
-        verifyJWT,
-        catchRequestError(async (req, res) => {
-            const id = req.body.id;
-            const tempId = JSON.stringify(id);
-
-            const [books] = await dbPoolPromise.execute(
-                `SELECT * FROM books WHERE JSON_CONTAINS(authors, '${tempId}')`
-            );
-            books.forEach(async (book) => {
-                const newAuthors = book.authors.filter(
-                    (author) => author !== id
-                );
-
-                await dbPoolPromise.execute(
-                    "UPDATE books SET authors = ? WHERE id = ?",
-                    [newAuthors, book.id]
-                );
-            });
-
-            await dbPoolPromise.execute("DELETE FROM authors WHERE id = ?", [
-                id
-            ]);
-
-            res.send("author deleted");
-        })
-    );
+    .post(verifyJWT, catchRequestError(author.findById))
+    .put(verifyJWT, catchRequestError(author.editById))
+    .delete(verifyJWT, catchRequestError(author.deleteById));
 
 module.exports = router;
